refactor: reduce cli.py to a 128-line registration facade (ADR-061 Phase 4)
`cli.py` goes from 1959 lines to 128. The `dump`/`compare` command bodies and
the shared CLI runtime moved into `abicheck/frontends/cli/`; what remains is
the Click root group, its `--version`/SIGTERM wiring, the tail-of-module
registration imports, and a lazy `__getattr__` keeping every historical
`abicheck.cli` spelling resolvable. That shim resolves through
`importlib.import_module` at *access* time -- a runtime call, not a static
import edge -- so the facade never grows a top-level dependency on the packages
that import back into it for `main`.

The interesting work was not the move. `frontends/` is a *migrated* package, so
`unclassified-import` applies to everything physically inside it, which forced
classifying the whole `cli_*` family -- and that surfaced 47 real direction
violations: the CLI reaching past the engine into `policy`, `compare` and
`extract`. They are closed, not suppressed. Each routes through a `workflows`
re-export surface, and the largest of them is Phase 4's own item 4 made
executable: `workflows/gate.py` is the one place a frontend gets its process
response, because three orthogonal axes feed one exit code and a frontend
importing them separately is free to fold two and forget the third.
`workflows/scan_config.py` is an owner rather than an alias -- it holds the
three functions `service_scan` previously imported *upward* for, which were the
last allowlist entries of their kind (15 -> 4 across Phases 3 and 4).

Two consequences are recorded rather than papered over, because both are easy
to miss and neither is invented here. A `monkeypatch.setattr` against a name
resolved through the shim rebinds nothing the real caller reads; and a
re-export *binds* at import time, so patching the original module does not
reach a caller coming through the facade. The suite was repointed at the
owners, and doing so found one patch target
(`abicheck.cli._detect_binary_format`) that had been inert *before* this
change -- `_normalize_binary_input` always resolved it through `cli_resolve`'s
own global, so the split surfaced a pre-existing vacuous patch rather than
creating one.

`abicheck.cli` is deliberately NOT added to `modules.yaml`'s `facades` list:
that gate's rule is narrower than the ADR's prose -- imports and inert
assignments only -- so any module declaring a Click root group fails it.
Widening it would weaken it for the pure re-export modules it was written for,
so the 150-line budget is pinned in `tests/test_cli_moved_surface.py`, which
also resolves every entry of the moved-name map and checks that the map covers
what the tree actually imports. A lazy shim is exactly the compatibility layer
that rots silently: a stale entry raises only when someone imports that one
name.

The import-cycle allowlist gains the three new modules on the same
rename-follows-member terms already recorded twice in that file: every edge
they have is an edge `cli` already had, and the loop still closes only through
the registration imports it always closed through.

`service.py` (1763 lines) is untouched and Phase 4's other half stays open; its
three large functions *are* the dump/compare implementation, and giving them a
`workflows` home is Phase 3's item 2, not a frontend translation.
