+++ abicheck/cli_scan_baseline.py
@@ -40,9 +40,10 @@ from typing import TYPE_CHECKING, Any
 
 import click
 
-from .buildsource.risk import RiskRules
 from .buildsource.scan_levels import EvidenceDepth, SourceMethod
 from .checker_policy import ADDITION_KINDS
+from .errors import SnapshotError
+from .workflows.scan_config import RiskRules
 
 if TYPE_CHECKING:
     from .environment_matrix import EnvironmentMatrix
@@ -54,29 +55,15 @@ if TYPE_CHECKING:
 def _public_provenance_set(
     headers: list[Path], public_header_dirs: list[Path]
 ) -> tuple[list[Path], list[Path]]:
-    """Build the ``(public_headers, public_header_dirs)`` provenance set for scan.
+    """CLI alias for ``workflows.scan_config.public_provenance_set``.
 
-    A directory boundary is what lets ``apply_provenance`` classify origins as
-    PUBLIC/INTERNAL (and so unlocks the leakage / RTTI / exported-vs-public
-    cross-checks, ADR-024). Directories come from ``--public-header-dir`` and from
-    any ``-H`` argument that is itself a directory; ``-H`` *file* arguments ride
-    along as explicit public headers.
-
-    A lone ``-H`` umbrella *file* with no directory does **not** activate
-    provenance: a single header cannot establish a public directory boundary
-    (the abicheck A1 finding), so we return empty sets and every origin stays
-    ``UNKNOWN`` — preserving the prior default-scan behaviour.
+    The rule moved to the engine in ADR-061 Phase 4 (``service_scan`` needed it
+    and had to import upward for it); this spelling stays because several call
+    sites and tests use it.
     """
-    dirs = list(public_header_dirs)
-    files: list[Path] = []
-    for h in headers:
-        if h.is_dir():
-            dirs.append(h)
-        else:
-            files.append(h)
-    if not dirs:
-        return [], []
-    return files, dirs
+    from .workflows.scan_config import public_provenance_set
+
+    return public_provenance_set(headers, public_header_dirs)
 
 
 def _expand_public_headers(headers: list[Path]) -> list[str]:
@@ -181,20 +168,18 @@ def _emit_estimate(
 
 
 def _load_risk_rules(path: Path | None) -> RiskRules:
-    """Load a ``risk_rules:`` profile from a YAML file, or the shipped default."""
-    if path is None:
-        return RiskRules.default()
-    import yaml  # hard dep (pyyaml); import out of the try so the except can name it
+    """CLI adapter over ``workflows.scan_config.load_risk_rules``.
+
+    Translates the engine's ``SnapshotError`` into a plain ``ClickException``
+    (**exit 1** -- operational, not a usage error: the flag was well-formed and
+    the file was not). Message unchanged from before the move.
+    """
+    from .workflows.scan_config import load_risk_rules
 
     try:
-        raw = yaml.safe_load(path.read_text(encoding="utf-8"))
-    except (OSError, ValueError, yaml.YAMLError) as exc:
-        # yaml.YAMLError (e.g. ParserError) is not a ValueError, so catch it
-        # explicitly — else malformed --risk-rules YAML escapes as a traceback
-        # through the installed console script (Codex review).
-        raise click.ClickException(f"cannot read --risk-rules {path}: {exc}") from exc
-    block = raw.get("risk_rules") if isinstance(raw, dict) else None
-    return RiskRules.from_dict(block if isinstance(block, dict) else raw)
+        return load_risk_rules(path)
+    except SnapshotError as exc:
+        raise click.ClickException(str(exc)) from exc
 
 
 #: Default cap on findings embedded in the ``scan --baseline`` summary so a
@@ -438,7 +423,7 @@ def _blocking_compatible_changes(diff: Any, blamed: set[str]) -> list[Any]:
     report, even though the gate itself (`_build_severity_json`, which does
     already pass `policy_file`) correctly named it as blocking.
     """
-    from .severity import classify_change_object
+    from .workflows.gate import classify_change_object
 
     kept: list[Any] = []
     for change in list(getattr(diff, "compatible", ()) or ()):
@@ -565,8 +550,8 @@ def _baseline_finding_dicts(
     with no way to tell *which* ``reclassify:`` rule produced it, unlike the
     compare/report path.
     """
-    from .finding_identity import report_canonical_finding_id, report_finding_id
     from .reporter import _reclassified_by_for_change
+    from .workflows.findings import report_canonical_finding_id, report_finding_id
 
     findings = []
     for c in changes:
@@ -673,7 +658,7 @@ def _baseline_is_native_library(path: Path) -> bool:
     name = path.name.lower()
     if name.endswith((".json", ".dump", ".tar.gz", ".tgz", ".xml")):
         return False
-    from .binary_utils import detect_binary_format
+    from .workflows.extraction import detect_binary_format
 
     if detect_binary_format(path) is not None:
         return True
@@ -943,7 +928,7 @@ def _baseline_summary(
     # separate `compare` invocation. `checker.compare` (reached through
     # `compare_snapshots` above) always attaches the result to *diff*, so
     # this is unconditional here too, exactly like `compare`'s report.
-    from .analysis_assurance import (
+    from .workflows.gate import (
         analysis_assurance_exit_contribution,
         analysis_assurance_report_dict,
     )
@@ -986,8 +971,8 @@ def _baseline_contract_block(diff: Any, resolved_config: Any) -> dict[str, Any]:
     context = context_block(diff)
     if context is None:
         return {}
-    from .contract_coverage_exit import coverage_exit_for_context
     from .contract_coverage_ledger import coverage_failures_for_context
+    from .workflows.gate import coverage_exit_for_context
 
     # The sibling unsuppressible ledger, on the same terms `compare`
     # reports it (plan Section 6.1) -- a coverage failure is not a
@@ -1209,7 +1194,7 @@ def _run_baseline_compare(
     # --require-complete-analysis.
     if requested_depth is not None:
         diff.requested_depth = requested_depth
-        from .analysis_assurance import compute_analysis_assurance
+        from .workflows.gate import compute_analysis_assurance
 
         diff.analysis_assurance = compute_analysis_assurance(
             diff,
@@ -1249,7 +1234,7 @@ def _run_baseline_compare(
     # that configured `exit_code_scheme="severity"` without ever resolving
     # a `sev_config` gets the identical legacy-scheme answer both places
     # agree on, rather than the resolver's severity branch's own assertion.
-    from .exit_decision import resolve_compare_exit_decision
+    from .workflows.gate import resolve_compare_exit_decision
 
     exit_scheme = (
         "severity"
@@ -1280,7 +1265,7 @@ def _run_baseline_compare(
     )
 
     from .cli_compare_helpers import _verdict_exit_code
-    from .contract_coverage_exit import fold_coverage_exit
+    from .workflows.gate import fold_coverage_exit
 
     verdict = diff.verdict.value
     # Mirrors `compare`'s own `_exit_with_severity_or_verdict` (cli.py):
@@ -1340,7 +1325,7 @@ def _run_baseline_compare(
     # folds both immediately in sequence so a caller cannot pick up one
     # orthogonal axis and forget the other. `0` contribution, and this is a
     # pure no-op, whenever the flag was not passed (default False).
-    from .analysis_assurance import (
+    from .workflows.gate import (
         assurance_floor_diagnostic,
         fold_analysis_assurance_exit,
     )
