+++ pyproject.toml
@@ -174,6 +174,11 @@ module = [
     "abicheck.cli_suggest",
     "abicheck.cli_surface",
     "abicheck.compat.cli",
+    # ADR-061 Phase 4 moved `cli.py`'s command bodies and shared runtime here;
+    # they carry the same click-decorator surface `cli` itself did.
+    "abicheck.frontends.cli.commands.compare",
+    "abicheck.frontends.cli.commands.dump",
+    "abicheck.frontends.cli.runtime",
 ]
 disallow_untyped_decorators = false
 disable_error_code = ["no-any-return", "misc"]
